refactor: rename flow_hooks to flowHooks
use js standards for naming, not ruby!
